+++ tsconfig.strict.json
@@ -66,7 +66,7 @@
     "./src/Platform/Hosted/Components/MeControl.test.tsx",
     "./src/Platform/Hosted/Components/MeControl.tsx",
     "./src/Platform/Hosted/Components/SignInButton.tsx",
-    "./src/Platform/Hosted/Helpers/PortalBackendClient.ts",
+    "./src/Platform/Hosted/PortalBackendConnectionStringClient.ts",
     "./src/Platform/Hosted/HostedUtils.test.ts",
     "./src/Platform/Hosted/HostedUtils.ts",
     "./src/Platform/Hosted/extractFeatures.test.ts",
